feat: Improve Workflow editor.
